+++ src/components/location-list-main.tsx
@@ -2,8 +2,8 @@ import classNames from 'classnames';
 import { AppRoute, CITIES } from './const';
 import { Link } from 'react-router-dom';
 import { useDispatch, useSelector } from 'react-redux';
-import { setCurrentCity } from '../reducer/cities/cities-slice.ts';
-import { RootState } from '../store.ts';
+import { setCurrentCity } from '../store/slices/cities-slice.ts';
+import { RootState } from '../store/index.ts';
 
 function LocationListMain(): JSX.Element {
   const dispatch = useDispatch();
