+++ src/index.tsx
@@ -2,7 +2,7 @@ import ReactDOM from 'react-dom/client';
 import App from './components/app';
 import './styles.css';
 import { Provider } from 'react-redux';
-import { store } from './store.ts';
+import { store } from './store/index.ts';
 
 const root = ReactDOM.createRoot(
   document.getElementById('root') as HTMLElement
