+++ src/services/token.ts
@@ -2,8 +2,7 @@ export const AUTH_TOKEN_KEY = 'six-cities';
 
 export type Token = string;
 
-const getToken = (): Token =>
-  localStorage.getItem(AUTH_TOKEN_KEY) || 'T2xpdmVyLmNvbm5lckBnbWFpbC5jb20=';
+const getToken = (): Token => localStorage.getItem(AUTH_TOKEN_KEY) || '';
 
 const saveToken = (token: Token): void =>
   localStorage.setItem(AUTH_TOKEN_KEY, token);
