+++ src/store/middleware/user-thunk.ts
@@ -1,0 +1,28 @@
+import { createAsyncThunk } from '@reduxjs/toolkit';
+import { AuthPayload, AuthResponse } from '../slices/user-slice';
+import { API } from '../../services/api';
+
+export const fetchLogin = createAsyncThunk(
+  'login/fetchLogin',
+  async (_, { rejectWithValue }) => {
+    try {
+      const { data } = await API.get<AuthResponse>('/login');
+      return data;
+    } catch (err) {
+      return rejectWithValue('Failed to load login');
+    }
+  }
+);
+
+export const tryAuth = createAsyncThunk(
+  'login/tryAuth',
+  async (payload: AuthPayload, { rejectWithValue }) => {
+    // Указываем payload с типом
+    try {
+      const { data } = await API.post<AuthResponse>('/login', payload);
+      return data;
+    } catch (err) {
+      return rejectWithValue('Failed to try auth');
+    }
+  }
+);
