+++ src/store/slices/cities-slice.ts
@@ -1,4 +1,4 @@
-import { createAsyncThunk, createSlice, PayloadAction } from '@reduxjs/toolkit';
+import { createSlice, PayloadAction } from '@reduxjs/toolkit';
 import {
   CurrentOfferType,
   OffersType,
@@ -6,7 +6,13 @@ import {
   UserInfo,
 } from '../../utils/type';
 import { CITIES, CitiesEnum, RequestStatus } from '../../components/const';
-import { API } from '../../services.ts/api';
+import {
+  fetchAllOffers,
+  fetchComments,
+  fetchCurrentOffer,
+  fetchNearbyOffers,
+  sendComment,
+} from '../middleware/cities-thunk';
 
 const DEFAULT_CITY =
   CITIES.find((item) => item.name === CitiesEnum.Paris) || CITIES[0];
@@ -53,76 +59,6 @@ const initialState: CitiesState = {
   nearbyOffers: [],
   comments: [],
 };
-
-// Создаем thunk для загрузки предложений
-export const fetchAllOffers = createAsyncThunk(
-  'offers/fetchAllOffers',
-  async (_, { extra, rejectWithValue }) => {
-    try {
-      const { data } = await (extra as typeof API).get<OffersType[]>('/offers');
-      return data;
-    } catch (err) {
-      return rejectWithValue('Failed to load offers');
-    }
-  }
-);
-
-export const fetchCurrentOffer = createAsyncThunk(
-  'offers/fetchCurrentOffer',
-  async (offerId: string, { extra, rejectWithValue }) => {
-    try {
-      const { data } = await (extra as typeof API).get<CurrentOfferType>(
-        `/offers/${offerId}`
-      );
-      return data;
-    } catch (err) {
-      return rejectWithValue('Failed to load currentOffer');
-    }
-  }
-);
-
-export const fetchNearbyOffers = createAsyncThunk(
-  'offers/fetchNearbyOffers',
-  async (offerId: string, { extra, rejectWithValue }) => {
-    try {
-      const { data } = await (extra as typeof API).get<OffersType[]>(
-        `/offers/${offerId}/nearby`
-      );
-      return data;
-    } catch (err) {
-      return rejectWithValue('Failed to load nearby offers');
-    }
-  }
-);
-
-export const fetchComments = createAsyncThunk(
-  'comments/fetchComments',
-  async (offerId: string, { extra, rejectWithValue }) => {
-    try {
-      const { data } = await (extra as typeof API).get<ReviewType[]>(
-        `/comments/${offerId}`
-      );
-      return data;
-    } catch (err) {
-      return rejectWithValue('Failed to load comments');
-    }
-  }
-);
-
-export const sendComment = createAsyncThunk(
-  'comments/sendComment',
-  async (payload: NewComment, { extra, rejectWithValue }) => {
-    try {
-      const { data } = await (extra as typeof API).post<ReviewType>(
-        `/comments/${payload.offerId}`,
-        { comment: payload.comment, rating: Number(payload.rating) }
-      );
-      return data;
-    } catch (err) {
-      return rejectWithValue('Failed to load new comment');
-    }
-  }
-);
 
 // хранилище
 const citiesSlice = createSlice({
